Refactor for separate explication pass and context
