+++ src/assembly/context.rs
@@ -7,8 +7,19 @@ use crate::assembly::ast::{
 use crate::ir;
 use debug_ignore::DebugIgnore;
 use std::collections::{HashMap, HashSet};
-use std::fmt::Debug;
-use std::hash::Hash;
+use crate::assembly::table::Table;
+
+#[derive(Debug)]
+pub struct Context {
+    pub ffi_type_table: Table<FFIType>,
+    pub local_type_table: Table<String>,
+    pub variable_map: HashMap<FuncletId, NodeTable>,
+    // where we currently are in the AST, using names
+    // optional cause we may not have started traversal
+    pub location: LocationNames,
+    pub funclet_indices: FuncletIndices,
+    pub function_classes: Table<FunctionClassId>
+}
 
 #[derive(Debug)]
 pub struct LocationNames {
@@ -18,59 +29,10 @@ pub struct LocationNames {
 }
 
 #[derive(Debug)]
-pub struct Table<T>
-where
-    T: Eq + Hash + Debug + Clone,
-{
-    values: HashSet<T>,
-    indices: Vec<T>,
-}
-
-#[derive(Debug)]
 pub struct NodeTable {
     // local names and return names such as [%out : i64] or whatever
     pub local: Table<NodeId>,
     pub returns: Table<NodeId>,
-}
-
-#[derive(Debug)]
-pub struct ValueFuncletConnections {
-    // stores connections of what refers to this value funclet
-    pub schedule_funclets: Vec<FuncletId>,
-    pub timeline_funclets: Vec<FuncletId>,
-    pub spatial_funclets: Vec<FuncletId>,
-}
-
-#[derive(Debug)]
-struct ValueExplicationInformation {
-    // explication locations are in the scheduling world
-    // maps from this node to the places it's been allocated on
-    scheduled_allocations: HashMap<FuncletId, NodeId>,
-
-    // indicates whether this operation has been written yet
-    // used primarily to add operations when needed
-    written: bool,
-}
-
-#[derive(Debug)]
-struct ValueFuncletData {
-    pub connections: ValueFuncletConnections,
-
-    // information about allocated value elements
-    explication_information: HashMap<NodeId, ValueExplicationInformation>,
-
-    // map from call index to output name for each call
-    call_outputs: HashMap<NodeId, HashMap<usize, NodeId>>,
-}
-
-#[derive(Debug)]
-struct ScheduleFuncletData {
-    // associated value funclet
-    value_funclet: FuncletId,
-    // map from the variables available to which node they are allocated
-    allocations: HashMap<NodeId, NodeId>,
-    // list of explication holes found, by index
-    explication_holes: Vec<usize>,
 }
 
 #[derive(Debug, Clone)]
@@ -115,98 +77,6 @@ pub struct FuncletIndices {
     funclet_kind_map: HashMap<String, FuncletLocation>,
 }
 
-#[derive(Debug)]
-struct MetaData {
-    name_index: usize,
-}
-
-#[derive(Debug)]
-pub struct Context {
-    pub ffi_type_table: Table<FFIType>,
-    pub local_type_table: Table<String>,
-    pub variable_map: HashMap<FuncletId, NodeTable>,
-    // where we currently are in the AST, using names
-    // optional cause we may not have started traversal
-    pub location: LocationNames,
-    pub funclet_indices: FuncletIndices,
-    pub function_classes: Table<FunctionClassId>,
-
-    // information found about a given value funclet
-    value_explication_data: HashMap<FuncletId, ValueFuncletData>,
-    // information found about a given schedule funclet
-    schedule_explication_data: HashMap<FuncletId, ScheduleFuncletData>,
-    meta_data: MetaData,
-    // connections from a given funclet
-}
-
-// a Table is basically a vector with no dupes
-impl<T> Table<T>
-where
-    T: Eq + Hash + Debug + Clone,
-{
-    pub fn new() -> Table<T> {
-        Table {
-            values: HashSet::new(),
-            indices: Vec::new(),
-        }
-    }
-
-    pub fn contains(&mut self, val: &T) -> bool {
-        self.values.contains(val)
-    }
-
-    pub fn dummy_push(&mut self, val: T) {
-        // Add unnamed element for indexing
-        self.indices.push(val);
-    }
-
-    pub fn push(&mut self, val: T) {
-        let msg = format!("Duplicate add of {:?}", val);
-        if !self.try_push(val) {
-            panic!(msg)
-        }
-    }
-
-    pub fn try_push(&mut self, val: T) -> bool {
-        self.indices.push(val.clone());
-        self.values.insert(val)
-    }
-
-    pub fn insert(&mut self, index: usize, val: T) {
-        if self.values.contains(&val) {
-            panic!("Duplicate add of {:?}", val)
-        }
-        self.values.insert(val.clone());
-        self.indices.insert(index, val);
-    }
-
-    pub fn get(&self, val: &T) -> Option<usize> {
-        // no need to actually check the Hashset, that's just to avoid dupes
-        for item in itertools::enumerate(&self.indices) {
-            if item.1 == val {
-                return Some(item.0);
-            }
-        }
-        return None;
-    }
-
-    pub fn get_index(&self, val: &T) -> Option<usize> {
-        self.get(val)
-    }
-
-    pub fn get_at_index(&self, index: usize) -> Option<&T> {
-        if index >= self.indices.len() {
-            None
-        } else {
-            Some(&self.indices[index])
-        }
-    }
-
-    pub fn len(&mut self) -> usize {
-        return self.indices.len();
-    }
-}
-
 impl LocationNames {
     pub fn new() -> LocationNames {
         LocationNames {
@@ -222,60 +92,6 @@ impl NodeTable {
             local: Table::new(),
             returns: Table::new(),
         }
-    }
-}
-
-impl ValueFuncletConnections {
-    pub fn new() -> ValueFuncletConnections {
-        ValueFuncletConnections {
-            schedule_funclets: Vec::new(),
-            timeline_funclets: Vec::new(),
-            spatial_funclets: Vec::new(),
-        }
-    }
-}
-
-impl ValueExplicationInformation {
-    pub fn new() -> ValueExplicationInformation {
-        ValueExplicationInformation {
-            scheduled_allocations: HashMap::new(),
-            written: false,
-        }
-    }
-}
-
-impl ValueFuncletData {
-    pub fn new() -> ValueFuncletData {
-        ValueFuncletData {
-            connections: ValueFuncletConnections::new(),
-            explication_information: HashMap::new(),
-            call_outputs: HashMap::new(),
-        }
-    }
-    pub fn allocate(
-        &mut self,
-        value_node: NodeId,
-        schedule_funclet: FuncletId,
-        schedule_node: NodeId,
-    ) {
-        self.explication_information
-            .entry(value_node)
-            .or_insert(ValueExplicationInformation::new())
-            .scheduled_allocations
-            .insert(schedule_funclet, schedule_node);
-    }
-}
-
-impl ScheduleFuncletData {
-    pub fn new(value_funclet: FuncletId) -> ScheduleFuncletData {
-        ScheduleFuncletData {
-            value_funclet,
-            allocations: HashMap::new(),
-            explication_holes: Vec::new(),
-        }
-    }
-    pub fn allocate(&mut self, schedule_node: NodeId, value_node: NodeId) {
-        self.allocations.insert(schedule_node, value_node);
     }
 }
 
@@ -322,22 +138,9 @@ impl FuncletIndices {
     }
 }
 
-impl MetaData {
-    pub fn new(name_index: usize) -> MetaData {
-        MetaData { name_index }
-    }
-    pub fn next_name(&mut self) -> String {
-        self.name_index += 1;
-        format!("~{}", self.name_index)
-    }
-}
-
 impl Context {
-    pub fn new(program: &ast::Program, name_index: usize) -> Context {
+    pub fn new(program: &ast::Program) -> Context {
         let mut context = Context {
-            value_explication_data: HashMap::new(),
-            schedule_explication_data: HashMap::new(),
-            meta_data: MetaData::new(name_index),
             ffi_type_table: Table::new(),
             local_type_table: Table::new(),
             funclet_indices: FuncletIndices::new(),
@@ -360,13 +163,6 @@ impl Context {
                     ast::TypeDecl::Local(t) => self.local_type_table.push(t.name.clone()),
                 },
                 ast::Declaration::Funclet(f) => {
-                    match f.kind {
-                        ir::FuncletKind::Value => {
-                            self.value_explication_data
-                                .insert(f.header.name.clone(), ValueFuncletData::new());
-                        }
-                        _ => {}
-                    };
                     self.funclet_indices
                         .insert(f.header.name.0.clone(), FuncletLocation::Local);
                     let mut node_table = NodeTable::new();
@@ -415,19 +211,6 @@ impl Context {
                 _ => {}
             }
         }
-    }
-
-    pub fn setup_schedule_data(&mut self, value_funclet: FuncletId) {
-        self.value_explication_data
-            .get_mut(&value_funclet)
-            .unwrap()
-            .connections
-            .schedule_funclets
-            .push(self.location.funclet_name.clone());
-        self.schedule_explication_data.insert(
-            self.location.funclet_name.clone(),
-            ScheduleFuncletData::new(value_funclet),
-        );
     }
 
     pub fn ffi_type_id(&self, name: &ast::FFIType) -> usize {
@@ -513,78 +296,5 @@ impl Context {
                 .clone(),
             node_id: self.remote_node_id(funclet, var),
         }
-    }
-
-    fn allocation_name(&mut self) -> String {
-        self.meta_data.name_index += 1;
-        format!("${}", self.meta_data.name_index)
-    }
-
-    pub fn add_allocation(&mut self, value_node: NodeId, schedule_node: NodeId) {
-        let schedule_funclet = self.location.funclet_name.clone();
-        let value_funclet = &self
-            .schedule_explication_data
-            .get(&schedule_funclet)
-            .unwrap()
-            .value_funclet;
-
-        self.variable_map
-            .get_mut(&schedule_funclet)
-            .unwrap()
-            .local
-            .try_push(schedule_node.clone());
-
-        // unwrap explicitly cause we assume funclet data are setup earlier
-        self.value_explication_data
-            .get_mut(value_funclet)
-            .unwrap()
-            .allocate(
-                value_node.clone(),
-                schedule_funclet.clone(),
-                schedule_node.clone(),
-            );
-
-        self.schedule_explication_data
-            .get_mut(&schedule_funclet)
-            .unwrap()
-            .allocate(schedule_node, value_node);
-    }
-
-    // get allocations of the associated value node
-    pub fn get_schedule_allocations(
-        &self,
-        funclet: &FuncletId,
-        node: &NodeId,
-    ) -> Option<&HashMap<FuncletId, NodeId>> {
-        self.value_explication_data.get(funclet).and_then(|f| {
-            f.explication_information
-                .get(node)
-                .map(|n| &n.scheduled_allocations)
-        })
-    }
-
-    pub fn get_current_schedule_allocation(&self, node: &NodeId) -> Option<&NodeId> {
-        self.get_current_value_funclet().as_ref().and_then(|vf| {
-            self.get_schedule_allocations(&vf, node)
-                .unwrap()
-                .get(&self.location.funclet_name)
-        })
-    }
-
-    // get what the associated schedule node is allocating
-    pub fn get_value_allocation(&self, funclet: &FuncletId, node: &NodeId) -> Option<&ast::NodeId> {
-        self.schedule_explication_data
-            .get(funclet)
-            .and_then(|f| f.allocations.get(node))
-    }
-
-    pub fn get_current_value_funclet(&self) -> Option<&FuncletId> {
-        self.schedule_explication_data
-            .get(&self.location.funclet_name)
-            .map(|f| &f.value_funclet)
-    }
-
-    pub fn next_name(&mut self) -> String {
-        self.meta_data.next_name()
     }
 }
