+++ src/assembly/explication/util.rs
@@ -1,0 +1,25 @@
+use crate::assembly::ast;
+use crate::assembly::ast::FFIType;
+use crate::assembly::ast::Hole;
+use crate::assembly::ast::{
+    ExternalFunctionId, FuncletId, FunctionClassId, NodeId, StorageTypeId, TypeId,
+};
+use crate::assembly::context;
+use crate::assembly::context::Context;
+use crate::assembly::parser;
+use crate::ir::ffi;
+use crate::{frontend, ir};
+use std::any::Any;
+use std::collections::HashMap;
+
+pub fn get_first<'a, T>(v: &'a Vec<T>, test: fn(&T) -> bool) -> Option<&'a T>
+where
+    T: Sized,
+{
+    for item in v {
+        if test(item) {
+            return Some(&item);
+        }
+    }
+    None
+}
