--- src/assembly/explication_util.rs
@@ -1,124 +1,0 @@
-use crate::assembly::ast;
-use crate::assembly::ast::FFIType;
-use crate::assembly::ast::Hole;
-use crate::assembly::ast::{
-    ExternalFunctionId, FuncletId, FunctionClassId, NodeId, StorageTypeId, TypeId,
-};
-use crate::assembly::context;
-use crate::assembly::context::Context;
-use crate::assembly::explication_explicator;
-use crate::assembly::parser;
-use crate::ir::ffi;
-use crate::{frontend, ir};
-use std::any::Any;
-use std::collections::HashMap;
-
-pub fn reject_hole<T>(h: Hole<T>) -> T {
-    match h {
-        Some(v) => v,
-        None => unreachable!("Unimplemented Hole"),
-    }
-}
-
-pub fn ffi_to_ffi(value: FFIType, context: &mut Context) -> ffi::Type {
-    fn box_map(b: Box<[FFIType]>, context: &mut Context) -> Box<[ffi::TypeId]> {
-        b.iter()
-            .map(|x| ffi::TypeId(context.ffi_type_id(x)))
-            .collect()
-    }
-    fn type_id(element_type: Box<FFIType>, context: &mut Context) -> ffi::TypeId {
-        ffi::TypeId(context.ffi_type_id(element_type.as_ref()))
-    }
-    match value {
-        FFIType::F32 => ffi::Type::F32,
-        FFIType::F64 => ffi::Type::F64,
-        FFIType::U8 => ffi::Type::U8,
-        FFIType::U16 => ffi::Type::U16,
-        FFIType::U32 => ffi::Type::U32,
-        FFIType::U64 => ffi::Type::U64,
-        FFIType::USize => ffi::Type::USize,
-        FFIType::I8 => ffi::Type::I8,
-        FFIType::I16 => ffi::Type::I16,
-        FFIType::I32 => ffi::Type::I32,
-        FFIType::I64 => ffi::Type::I64,
-        FFIType::Array {
-            element_type,
-            length,
-        } => ffi::Type::Array {
-            element_type: type_id(element_type, context),
-            length,
-        },
-        FFIType::ErasedLengthArray(element_type) => ffi::Type::ErasedLengthArray {
-            element_type: type_id(element_type, context),
-        },
-        FFIType::Struct {
-            fields,
-            byte_alignment,
-            byte_size,
-        } => todo!(),
-        FFIType::Tuple(element_types) => ffi::Type::Tuple {
-            fields: box_map(element_types.into_boxed_slice(), context),
-        },
-        FFIType::ConstRef(element_type) => ffi::Type::ConstRef {
-            element_type: type_id(element_type, context),
-        },
-        FFIType::MutRef(element_type) => ffi::Type::MutRef {
-            element_type: type_id(element_type, context),
-        },
-        FFIType::ConstSlice(element_type) => ffi::Type::ConstSlice {
-            element_type: type_id(element_type, context),
-        },
-        FFIType::MutSlice(element_type) => ffi::Type::MutSlice {
-            element_type: type_id(element_type, context),
-        },
-        FFIType::GpuBufferRef(element_type) => ffi::Type::GpuBufferRef {
-            element_type: type_id(element_type, context),
-        },
-        FFIType::GpuBufferSlice(element_type) => ffi::Type::GpuBufferSlice {
-            element_type: type_id(element_type, context),
-        },
-        FFIType::GpuBufferAllocator => ffi::Type::GpuBufferAllocator,
-        FFIType::CpuBufferAllocator => ffi::Type::CpuBufferAllocator,
-        FFIType::CpuBufferRef(element_type) => ffi::Type::CpuBufferRef {
-            element_type: type_id(element_type, context),
-        },
-    }
-}
-
-pub fn remote_conversion(remote: &ast::RemoteNodeId, context: &Context) -> ir::RemoteNodeId {
-    remote_location_conversion(
-        &context::LocationNames {
-            funclet_name: remote.funclet_name.clone().unwrap(),
-            node_name: remote.node_name.clone().unwrap(),
-        },
-        context,
-    )
-}
-
-pub fn remote_location_conversion(
-    remote: &context::LocationNames,
-    context: &Context,
-) -> ir::RemoteNodeId {
-    ir::RemoteNodeId {
-        funclet_id: context
-            .funclet_indices
-            .get_funclet(&remote.funclet_name.0)
-            .unwrap()
-            .clone(),
-        node_id: context
-            .remote_node_id(&remote.funclet_name, &remote.node_name)
-            .clone(),
-    }
-}
-
-pub fn get_first<'a, T>(v: &'a Vec<T>, test: fn(&T) -> bool) -> Option<&'a T>
-where
-    T: Sized,
-{
-    for item in v {
-        if test(item) {
-            return Some(&item);
-        }
-    }
-    None
-}
