+++ src/assembly/mod.rs
@@ -5,5 +5,4 @@ pub mod ast;
 
 mod context;
 mod explication;
-mod explication_explicator;
-mod explication_util;
+mod table;
